chore: moved route-matcher.ts to ext folder
