+++ src/ext/routing.ts
@@ -1,5 +1,5 @@
 import { allTags, CTag } from '../tag.js';
-import { RouteMatcher, routeMatcher } from '../route-matcher.js';
+import { RouteMatcher, routeMatcher } from './route-matcher.js';
 import { onLifecycle } from '../lifecycle.js';
 
 const { div, a } = allTags;
